tests: replace jest with tap
