+++ lib/plugins/fs-routes/__test__/routes.test.js
@@ -1,9 +1,10 @@
 const path = require('path')
+const tap = require('tap')
 const fs = require('fs-extra')
 const { collectRoutes } = require('../routes-template')
 
 function testCollectRoutes(dir, options) {
-  test(`Collect routes in ${dir}`, async () => {
+  tap.test(`Collect routes in ${dir}`, async () => {
     const pagesDir = path.resolve(__dirname, dir)
     const routes = await collectRoutes(
       {
@@ -24,7 +25,7 @@ function testCollectRoutes(dir, options) {
         JSON.stringify(routes, null, 2) + '\n'
       )
     } else {
-      expect(routes).toEqual(await fs.readJson(expectedRoutesFile))
+      tap.same(routes, await fs.readJson(expectedRoutesFile))
     }
   })
 }
@@ -43,10 +44,11 @@ testCollectRoutes('special/custom-base-path', {
   basePath: '/some'
 })
 
-test('Return empty routes for non-existent directory', async () => {
-  expect(
+tap.test('Return empty routes for non-existent directory', async () => {
+  tap.same(
     await collectRoutes({
       dir: path.resolve(__dirname, 'no-such-path')
-    })
-  ).toEqual([])
+    }),
+    []
+  )
 })
