+++ lib/validateConfig.test.js
@@ -1,0 +1,14 @@
+const tap = require('tap')
+const validateConfig = require('./validateConfig')
+
+tap.doesNotThrow(() => {
+  const res = validateConfig({
+    entry: 'hi',
+    server: {
+      port: 7000
+    }
+  })
+  if (res.error) {
+    throw res.error
+  }
+})
