+++ package.json
@@ -9,16 +9,18 @@
   "bin": "bin/cli.js",
   "main": "lib/index.js",
   "scripts": {
-    "test": "xo && jest",
+    "test": "npm run lint && npm run tap",
+    "tap": "tap 'lib/**/*.test.js'",
     "lint": "xo",
     "postinstall": "node -e \"console.log('\\u001b[35m\\u001b[1mLove Ream? You can now donate to support the author:\\u001b[22m\\u001b[39m\\n> \\u001b[34mhttps://patreon.com/egoist\\u001b[0m')\""
   },
   "files": [
     "lib",
-    "!**/__test__/**",
     "bin",
     "app",
-    "webpack.config.js"
+    "webpack.config.js",
+    "!**/__test__/**",
+    "!**/*.test.js"
   ],
   "engines": {
     "node": ">=8",
@@ -120,11 +122,11 @@
     "eslint-plugin-jest": "^21.17.0",
     "eslint-plugin-prettier": "^2.6.0",
     "husky": "^1.0.0-rc.6",
-    "jest": "^23.1.0",
     "lerna": "^2.11.0",
     "lint-staged": "^7.1.0",
     "prettier": "^1.12.1",
     "semantic-release": "^15.9.17",
+    "tap": "^12.0.1",
     "vuex": "^3.0.1",
     "xo": "^0.20.1"
   },
